Create README.md
The main thing that I did in this version of my assignment is that I changed the picture to be plastic in the ocean and fish swimming in it. I went with this approach because I think that it’s better suited for the message that I am trying to get across, which is to limit plastic use which ultimately helps limit ocean pollution. I don’t think I am quite done it yet, I am thinking about adding something extra to it but I don’t know what that is at the moment.